% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commoncrawl.R
\name{cc_warc}
\alias{cc_warc}
\title{Provides WARC paths for commoncrawl.org}
\usage{
cc_warc(start, end = start)
}
\arguments{
\item{start}{The first path to retrieve.}

\item{end}{The last path to retrieve.}
}
\description{
Provides WARC paths for commoncrawl.org. To be used with
\code{spark_read_warc}.
}
\examples{

cc_warc(1)
cc_warc(2, 3)

}

