% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{rcpp_read_warc_sample}
\alias{rcpp_read_warc_sample}
\title{Loads the sample warc file in Rcpp}
\usage{
rcpp_read_warc_sample(filter = "", include = "")
}
\arguments{
\item{filter}{A regular expression used to filter to each warc entry
efficiently by running native code using \code{Rcpp}.}

\item{include}{A regular expression used to keep only matching lines
efficiently by running native code using \code{Rcpp}.}
}
\description{
Loads the sample warc file in Rcpp
}
