% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkwarc.R
\name{spark_rcpp_read_warc}
\alias{spark_rcpp_read_warc}
\title{Reads a WARC File into using Rcpp}
\usage{
spark_rcpp_read_warc(path, match_warc, match_line)
}
\arguments{
\item{path}{The path to the file. Needs to be accessible from the cluster.
Supports the \samp{"hdfs://"}, \samp{"s3n://"} and \samp{"file://"} protocols.}

\item{match_warc}{include only warc files mathcing this character string.}

\item{match_line}{include only lines mathcing this character string.}
}
\description{
Reads a WARC (Web ARChive) file using Rcpp.
}
