% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkwarc.R
\name{spark_read_warc}
\alias{spark_read_warc}
\title{Reads a WARC File into Apache Spark}
\usage{
spark_read_warc(sc, path, partitions = 10, group = FALSE, ...)
}
\arguments{
\item{sc}{An active \code{spark_connection}.}

\item{path}{The path to the warc file.}

\item{partitions}{Number of partitions to use for loaded warc.}

\item{group}{\code{TRUE} to group by warc segment.}

\item{...}{Additional arguments reserved for future use.}
}
\description{
Reads a WARC (Web ARChive) file into Apache Spark using sparklyr.
}
\examples{

library(sparklyr)
sc <- spark_connect(master = "spark://HOST:PORT")
df <- spark_read_warc(
  sc,
  system.file("samples/sample.warc", package = "sparkwarc")
)

spark_disconnect(sc)

}

