% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{spark_read_warc_sample}
\alias{spark_read_warc_sample}
\title{Loads the sample warc file}
\usage{
spark_read_warc_sample(filter = "")
}
\arguments{
\item{filter}{A regular expression that is applied to each warc entry
efficiently by running native code using \code{Rcpp}.}
}
\description{
Loads the sample warc file
}
